Rewrite `eraseOp` to prove it more easily, and add GetSet lemmas for the helpers (#52)
Using recursion will allow us to more easily prove `eraseOp` correct,
rather than using the `for` syntax which doesn't have any
lemmas/theorems.

The Get/Set lemmas for `detachOperands` and `detachOperands.loop` do not
contain all the usual ones, as it is quite hard to state most of these
lemmas, as they require to reason way more about the def-use chain. We
should instead use high-level lemmas for that.
For instance, it is not really useful to know what `(OpOperandPtr.get!
(detachOperands ...)).nextUse` is, since we only want to know that 1)
the def-use chain is well-formed, and 2) what are the uses or a
particular definition.